code_feedback/requirements.txt
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.9.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
bert-score==0.3.13
bitsandbytes==0.46.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
comm==0.2.2
contourpy==1.3.2
cut-cross-entropy==25.1.1
cycler==0.12.1
datasets==3.6.0
debugpy==1.8.14
decorator==5.2.1
diffusers==0.34.0
dill==0.3.8
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.3.0
gguf==0.17.1
h11==0.16.0
hf-xet==1.1.3
hf_transfer==0.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
langchain-core==0.3.65
langchain-ollama==0.3.3
langsmith==0.3.45
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgspec==0.19.0
multidict==6.6.3
multiprocess==0.70.16
narwhals==1.44.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
ollama==0.5.1
openai==1.55.3
orjson==3.10.18
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.16.0
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
plotly==6.2.0
pluggy==1.6.0
prometheus_client==0.20.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf>=3.20.0,<4.0.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==4.1.0
sentencepiece==0.2.0
shellingham==1.5.4
shtab==1.7.2
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.2
tomli==2.2.1
torch==2.7.0
torchvision==0.22.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.1
trl==0.19.0
typeguard==4.4.4
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.14.0
tyro==0.9.26
tzdata==2025.2
umap-learn==0.5.7
urllib3==2.4.0
wcwidth==0.2.13
wget==3.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0